refactored common modules
