+++ src/common/models/settings.ts
@@ -1,4 +1,4 @@
-export interface Settings {
+export interface SettingsModel {
   _id: string;
   currentTheme: string;
 }
