--- src/common/utils/isWindows.ts
@@ -1,4 +1,0 @@
-// because who needs `is-windows`
-const isWindows = () => process.platform === 'win32';
-
-export default isWindows;
